refactor(app-headless-cms): convert positional parameters to objects
