+++ packages/app-headless-cms/src/admin/plugins/fieldValidators/minLength.tsx
@@ -14,7 +14,7 @@ const plugin: CmsModelFieldValidatorPlugin = {
         description: "Entered value must not be shorter than the provided min length.",
         defaultMessage: "Value is too short.",
         variables: [{ name: "value", description: "This is the minimum allowed length." }],
-        getVariableValues: validator => {
+        getVariableValues: ({ validator }) => {
             return { value: validator.settings.value };
         },
         renderSettings(config) {
@@ -35,7 +35,7 @@ const plugin: CmsModelFieldValidatorPlugin = {
                 </Grid>
             );
         },
-        validate: async (value, validator) => {
+        validate: async (value, { validator }) => {
             const minLengthValue = validator.settings.value;
             if (typeof minLengthValue === "undefined") {
                 return true;
