+++ packages/app-headless-cms/src/admin/plugins/fieldValidators/unique.tsx
@@ -10,6 +10,9 @@ export default (): CmsModelFieldValidatorPlugin => {
             description: "You won't be able to submit the form if this field is not unique",
             defaultMessage: "Value must be unique.",
             validate: async () => {
+                /**
+                 * We let this validator be executed on the API side.
+                 */
                 return true;
             }
         }
